Commit the content of review 3857 to this branch.
The content of this changeset ensures that channels that
move from a Stasis bridge to a non-Stasis bridge are booted
from Stasis in the process. This means that after leaving
the Stasis bridge, a ChannelLeftBridge event is sent, followed
by a StasisEnd event. No futher events are issued for the
channel since the application is then unsubscribed from
that channel's events.