Address the rest of the review feedback.
The biggest change here was the one getting rid of the 2048 byte stack
allocation when creating an RLMI body. Now, instead of that, I store the
XML directly in the pjsip_msg_body() and provide a callback for it to
be converted to text.

In doing so, I changed the multiple #defines of the XML prolog length
in body generators to be defined in a header instead.