Make sure we use the proper address
Now, this is a situation where the IP address in the registry object does not
match the one actually used for the REGISTER transaction. We propably should
update the registry object as well. I am surprised that we look up the DNS
address so many times for a single REGISTER. With round robin DNS, the address
keeps changing, which cause bad things to happen.