astobj2: Correct REF_DEBUG false leak report
When ao2_callback is run with OBJ_MULTIPLE and not OBJ_NODATA
it allocates a temporary container in a way that does not
record REF_DEBUG log entries.  This changes that container
to correctly record unref's when the container is freed. 

ASTERISK-24390 #close
Reported by: Corey Farrell
Review: https://reviewboard.asterisk.org/r/4047/